+++ inc/extras-penguin-gold.php
@@ -54,15 +54,10 @@ add_action( 'stylesheet_uri', 'penguin_stylesheet_uri', 10, 2 );
 function penguin_gold_add_search_box($items, $args) {
 	$menusearch = get_theme_mod( 'menu-search' );
 	if ( 'primary' == $args->theme_location && 1 == $menusearch ) {
-		ob_start();
-		get_search_form();
-		$searchform = ob_get_contents();
-		ob_end_clean();
-
-		$items .= '<li class="menu-search">';
-		$items .= '<svg version="1.1" aria-hidden="true" class="penguin-icon-search">';
-		$items .= '<use xlink:href="' . esc_url( get_template_directory_uri() ) . '/icons.svg#penguin-icon-search"></use>';
-		$items .= '</svg>' . $searchform . '</li>';
+		$items .= '<li class="menu-search">
+			<svg version="1.1" aria-hidden="true" class="penguin-icon-search">
+			<use xlink:href="' . esc_url( get_template_directory_uri() ) . '/icons.svg#penguin-icon-search"></use>
+			</svg>' . get_search_form( false ) . '</li>';
 	}
 	return $items;
 }